+++ test/matchers/associations_test.rb
@@ -8,6 +8,8 @@ describe "Associations" do
     it { must have_many(:pets).of_type(Pet) }
     it { must have_one(:account).of_type(Account) }
     it { must have_and_belong_to_many(:friends).of_type(Person) }
+
+    it { must embed_one(:profile) }
   end
 
   describe Pet do